fix camera bugs of black incomplete frames

+++ hw/hw_gate.py
@@ -27,7 +27,7 @@ def motor_cmnd_to_DShot(cmnds):
         zero_val = 1048
         if np.sign(cmd) == -1:
             zero_val = 48
-        cmd_dshot = (zero_val + min(max(round(cmd*999), 0), 999)) << 1
+        cmd_dshot = int(zero_val + min(max(round(cmd*999), 0), 999)) << 1
         csum = (cmd_dshot ^ (cmd_dshot >> 4) ^ (cmd_dshot >> 8)) & 0xf
         dshot_msgs[idx] = cmd_dshot << 4 | csum
